add tests cont.
